+++ cli/commands/metadata_diff_test.go
@@ -13,19 +13,19 @@ import (
 
 var _ = Describe("hasura metadata diff", func() {
 
-	var dirName string
+	var projectDirectory string
 	var teardown func()
 	BeforeEach(func() {
-		dirName = testutil.RandDirName()
+		projectDirectory = testutil.RandDirName()
 		hgeEndPort, teardownHGE := testutil.StartHasura(GinkgoT(), testutil.HasuraDockerImage)
 		hgeEndpoint := fmt.Sprintf("http://0.0.0.0:%s", hgeEndPort)
 		testutil.RunCommandAndSucceed(testutil.CmdOpts{
-			Args: []string{"init", dirName},
+			Args: []string{"init", projectDirectory},
 		})
-		editEndpointInConfig(filepath.Join(dirName, defaultConfigFilename), hgeEndpoint)
+		editEndpointInConfig(filepath.Join(projectDirectory, defaultConfigFilename), hgeEndpoint)
 
 		teardown = func() {
-			os.RemoveAll(dirName)
+			os.RemoveAll(projectDirectory)
 			teardownHGE()
 		}
 	})
@@ -38,7 +38,7 @@ var _ = Describe("hasura metadata diff", func() {
 		It("should output diff between metadata on server and local project", func() {
 			session := testutil.Hasura(testutil.CmdOpts{
 				Args:             []string{"metadata", "diff"},
-				WorkingDirectory: dirName,
+				WorkingDirectory: projectDirectory,
 			})
 			Eventually(session, timeout).Should(Exit(0))
 			stdout := session.Out.Contents()
@@ -47,6 +47,36 @@ var _ = Describe("hasura metadata diff", func() {
 			Expect(stdout).Should(ContainSubstring("-  kind: postgres"))
 			Expect(stdout).Should(ContainSubstring("-  name: default"))
 			Expect(stdout).Should(ContainSubstring("-  tables: []"))
+
+			editMetadataFileInConfig(filepath.Join(projectDirectory, defaultConfigFilename), "metadata.yaml")
+			session = testutil.Hasura(testutil.CmdOpts{
+				Args:             []string{"metadata", "diff"},
+				WorkingDirectory: projectDirectory,
+			})
+			Eventually(session, timeout).Should(Exit(0))
+			stdout = session.Out.Contents()
+			Expect(stdout).Should(ContainSubstring("-sources:"))
+			Expect(stdout).Should(ContainSubstring("-  kind: postgres"))
+			Expect(stdout).Should(ContainSubstring("-- name: default"))
+			Expect(stdout).Should(ContainSubstring("-  tables: []"))
+
+			editMetadataFileInConfig(filepath.Join(projectDirectory, defaultConfigFilename), "metadata.json")
+			session = testutil.Hasura(testutil.CmdOpts{
+				Args:             []string{"metadata", "diff"},
+				WorkingDirectory: projectDirectory,
+			})
+			Eventually(session, timeout).Should(Exit(0))
+			stdout = session.Out.Contents()
+			want := `-  "version": 3,
+-  "sources": [
+-    {
+-      "name": "default",
+-      "kind": "postgres",
+-      "tables": [],
+-      "configuration": {
+`
+
+			Expect(stdout).Should(ContainSubstring(want))
 		})
 	})
 })