+++ cli/commands/metadata_export.go
@@ -4,7 +4,6 @@ import (
 	"io/ioutil"
 
 	"github.com/hasura/graphql-engine/cli/v2"
-	"github.com/hasura/graphql-engine/cli/v2/internal/projectmetadata"
 	"github.com/pkg/errors"
 	"github.com/spf13/cobra"
 )
@@ -33,9 +32,16 @@ func newMetadataExportCmd(ec *cli.ExecutionContext) *cobra.Command {
   hasura metadata export --endpoint "<endpoint>"`,
 		SilenceUsage: true,
 		RunE: func(cmd *cobra.Command, args []string) error {
+			if len(opts.output) == 0 {
+				ec.Spin("Exporting metadata...")
+			}
 			err := opts.Run()
+			ec.Spinner.Stop()
 			if err != nil {
 				return err
+			}
+			if len(opts.output) == 0 {
+				ec.Logger.Info("Metadata exported")
 			}
 			return nil
 		},
@@ -43,7 +49,7 @@ func newMetadataExportCmd(ec *cli.ExecutionContext) *cobra.Command {
 	}
 
 	f := metadataExportCmd.Flags()
-	f.StringVarP(&opts.output, "output", "o", "", `specify an output format for exported metadata (note: this won't modify project metadata) Allowed values: json, yaml")`)
+	f.StringVarP(&opts.output, "output", "o", "", `write metadata to standard output in given format for exported metadata (note: this won't modify project metadata) Allowed values: json, yaml")`)
 
 	return metadataExportCmd
 }
@@ -58,19 +64,7 @@ func (o *MetadataExportOptions) Run() error {
 	if len(o.output) != 0 {
 		return getMetadataFromServerAndWriteToStdoutByFormat(o.EC, rawOutputFormat(o.output))
 	}
-	o.EC.Spin("Exporting metadata...")
-	metadataHandler := projectmetadata.NewHandlerFromEC(o.EC)
-	files, err := metadataHandler.ExportMetadata()
-	o.EC.Spinner.Stop()
-	if err != nil {
-		return errors.Wrap(err, "failed to export metadata")
-	}
-	err = metadataHandler.WriteMetadata(files)
-	if err != nil {
-		return errors.Wrap(err, "cannot write metadata to project")
-	}
-	o.EC.Logger.Info("Metadata exported")
-	return nil
+	return getMetadataModeHandler(o.EC.MetadataMode).Export(o)
 }
 
 func getMetadataFromServerAndWriteToStdoutByFormat(ec *cli.ExecutionContext, format rawOutputFormat) error {