+++ console/src/components/Services/Data/DataActions.js
@@ -381,6 +381,9 @@ const fetchDataInit = (source, driver) => (dispatch, getState) => {
   );
 };
 
+/**
+ * @param {string| string[]} [schema=null]
+ */
 const fetchFunctionInit = (schema = null) => (dispatch, getState) => {
   const url = Endpoints.query;
   const source = getState().tables.currentDataSource;